chore: refactor-oc-table-header-to-setup-script
