+++ packages/design-system/src/components/OcTableHeader/OcTableHeader.vue
@@ -4,9 +4,8 @@
   </thead>
 </template>
 
-<script lang="ts">
-import { defineComponent } from 'vue'
-export default defineComponent({
+<script lang="ts" setup>
+defineOptions({
   name: 'OcThead',
   status: 'ready',
   release: '2.1.0'
