+++ packages/design-system/src/components/index.ts
@@ -44,7 +44,7 @@ export { default as OcTableBody } from './OcTableBody/OcTableBody.vue'
 export { default as OcTableCellData } from './OcTableCellData/OcTableCellData.vue'
 export { default as OcTableFooter } from './OcTableFooter/OcTableFooter.vue'
 export { default as OcTableCellHead } from './OcTableCellHead/OcTableCellHead.vue'
-export { default as OcTableHeader } from './_OcTableHeader/_OcTableHeader.vue'
+export { default as OcTableHeader } from './OcTableHeader/OcTableHeader.vue'
 export { default as OcTableRow } from './_OcTableRow/_OcTableRow.vue'
 export { default as OcTableSimple } from './OcTableSimple/OcTableSimple.vue'
 export { default as OcTag } from './OcTag/OcTag.vue'
